Style improvements
